move solver files to separate directory
